requirements.txt
torch==2.1.0
torchvision
sentence-transformers==2.2.2
onnx==1.14.1
onnxconverter-common==1.13.0
onnxruntime-gpu==1.16.1
transformers==4.33.3
tokenizers==0.13.3
tqdm==4.66.1
numpy==1.26.1
scikit-learn==1.3.2
scipy==1.11.3
nltk==3.8.1
sentencepiece==0.1.99
huggingface-hub==0.18.0